pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "multi-agent-immuno"
version = "0.1.0"
description = "multi-agent-immuno project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "langchain-community>=0.3.19",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.8",
    "langgraph>=0.3.5",
    "readabilipy>=0.3.0",
    "python-dotenv>=1.0.1",
    "socksio>=1.0.0",
    "markdownify>=1.1.0",
    "browser-use>=0.1.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "sse-starlette>=1.6.5",
    "pandas>=2.2.3",
    "numpy>=2.2.3",
    "yfinance>=0.2.54",
    "langchain-deepseek>=0.1.2",
    "pydantic-ai-slim[openai]>=0.2.15",
    "crawl4ai>=0.6.3",
    "crewai>=0.130.0",
    "openai>=1.84.0",
    "matplotlib>=3.10.3",
    "langchain-sandbox>=0.0.6",
    "mcp>=1.9.4",
    "langgraph-cli[inmem]>=0.3.3",
    "langgraph-api>=0.2.54",
    "scipy>=1.15.3",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "plotly>=6.1.2",
    "statsmodels>=0.14.4",
    "langchain-tavily>=0.2.4",
    "openevals>=0.1.0",
    "langsmith>=0.3.45",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.2.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/build/
'''

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.0.0",
]
